updated test cases and jwt util
