+++ libs/jarc-auth-lib/src/main/java/com/sourcefuse/jarc/authlib/providers/JwtTokenDecryptProvider.java
@@ -1,13 +1,7 @@
 package com.sourcefuse.jarc.authlib.providers;
 
-import java.security.Key;
-import com.sourcefuse.jarc.authlib.Utils;
-import io.jsonwebtoken.io.Decoders;
-import io.jsonwebtoken.security.Keys;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.http.HttpStatus;
-import org.springframework.stereotype.Component;
-
+import com.sourcefuse.jarc.authlib.utils.JwtUtils;
+import com.sourcefuse.jarc.authlib.utils.Utils;
 import com.sourcefuse.jarc.core.constants.AuthConstants;
 import com.sourcefuse.jarc.core.exception.CommonRuntimeException;
 import com.sourcefuse.jarc.core.models.session.CurrentUser;
@@ -16,9 +10,11 @@ import io.jsonwebtoken.ExpiredJwtException;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.MalformedJwtException;
 import io.jsonwebtoken.UnsupportedJwtException;
-
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.http.HttpStatus;
+import org.springframework.stereotype.Component;
 
 @RequiredArgsConstructor
 @Component
@@ -27,15 +23,12 @@ public class JwtTokenDecryptProvider {
 
   @Value("${app.jwt-secret}")
   private String jwtSecret;
-  private Key key() {
-    return Keys.hmacShaKeyFor(Decoders.BASE64.decode(jwtSecret));
-  }
 
   public CurrentUser getUserDetails(String token) {
     try {
       Claims claims = Jwts
         .parserBuilder()
-        .setSigningKey(key())
+        .setSigningKey(JwtUtils.key(jwtSecret))
         .build()
         .parseClaimsJws(token)
         .getBody();
