+++ libs/jarc-auth-lib/src/main/java/com/sourcefuse/jarc/authlib/utils/JwtUtils.java
@@ -1,8 +1,5 @@
-package com.sourcefuse.jarc.authlib;
+package com.sourcefuse.jarc.authlib.utils;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.SerializationFeature;
-import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import com.sourcefuse.jarc.authlib.security.CustomJwtBuilder;
 import com.sourcefuse.jarc.core.constants.AuthConstants;
 import com.sourcefuse.jarc.core.models.session.CurrentUser;
@@ -10,21 +7,12 @@ import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.io.Decoders;
 import io.jsonwebtoken.security.Keys;
-
 import java.security.Key;
 import java.util.Date;
-import org.springframework.http.converter.json.Jackson2ObjectMapperBuilder;
 
-public final class Utils {
+public final class JwtUtils {
 
-  private Utils() {}
-
-  public static ObjectMapper getObjectMapperInstance() {
-    return new Jackson2ObjectMapperBuilder()
-      .featuresToDisable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS)
-      .modulesToInstall(new JavaTimeModule())
-      .build();
-  }
+  private JwtUtils() {}
 
   public static String generateAccessToken(
     String jwtSecret,
@@ -44,7 +32,7 @@ public final class Utils {
       .compact();
   }
 
-  private static Key key(String jwtSecret) {
+  public static Key key(String jwtSecret) {
     return Keys.hmacShaKeyFor(Decoders.BASE64.decode(jwtSecret));
   }
 }
