+++ libs/jarc-auth-lib/src/main/java/com/sourcefuse/jarc/authlib/utils/Utils.java
@@ -1,0 +1,18 @@
+package com.sourcefuse.jarc.authlib.utils;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.SerializationFeature;
+import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
+import org.springframework.http.converter.json.Jackson2ObjectMapperBuilder;
+
+public final class Utils {
+
+  private Utils() {}
+
+  public static ObjectMapper getObjectMapperInstance() {
+    return new Jackson2ObjectMapperBuilder()
+      .featuresToDisable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS)
+      .modulesToInstall(new JavaTimeModule())
+      .build();
+  }
+}
