+++ services/notification-service/src/test/resources/application.properties
@@ -1,0 +1,3 @@
+
+app.jwt-secret= daf66e01593f61a15b857cf433aae03a005812b31234e149036bcc8dee755dbb
+app.app-jwt-expiration-milliseconds= 3600000
